+++ src/transmetteurs/TransmetteurAnalogiqueParfaitMulti.java
@@ -1,14 +1,15 @@
 package transmetteurs;
 
 /**
- * Classe d'un composant qui transmet des informations de type Float avec des trajets multiples mais sans défaut.
+ * Classe d'un composant qui transmet des informations de type Double avec des
+ * trajets multiples mais sans défaut.
  *
  * @author Antoine GIRARD
  * @author Cédric HERZOG
  */
 public class TransmetteurAnalogiqueParfaitMulti extends TransmetteurAnalogiqueBruiteMulti {
 
-    public TransmetteurAnalogiqueParfaitMulti(Integer[] dt, Float[] ar) throws Exception {
+    public TransmetteurAnalogiqueParfaitMulti(Integer[] dt, Double[] ar) throws Exception {
         super(dt, ar, null);
     }
